nutrichain_restored/requirements.txt
flask>=2.0.0
pycryptodome>=3.15.0
